+++ vtm/resources/assets/shaders/linetex_layer_tex.glsl
@@ -40,13 +40,13 @@ uniform vec4 u_bgcolor;
 uniform float u_pwidth;
 varying vec2 v_st;
 uniform sampler2D tex;
-uniform float u_mode;
+uniform int u_mode;
 void
 main(){
-  if (u_mode >= 1.0) {
+  if (u_mode >= 1) {
     /* Dash array or texture */
     float step = 2.0;
-    if (u_mode == 2.0) { // dashed texture
+    if (u_mode == 2) { // dashed texture
       step = 1.0;
     }
     // use lineLength mod texture step (mod is always positive)
